move legacy plotly/colors.py to plotly/colors/__init__.py
